new ant tasks: package-lib, packge-examples
If you are modifying the source code and testing it with 
a new example or your own project the following ant tasks 
which complete in just a couple of seconds may be helpful.

ant package-lib

Create the "dist" directory and populate the lib directory with scenejs.js, scenejs.min.js, and the plugins and utils directories.

ant packge-examples

Create the "dist" directory and populate the lib directory and the examples directory.