Use VAO on the refactored geometry+morphGeometry chunk
The VAO reference is stored on the geometry chunk, which now handles
both morph geometry and regular geometry. If the interleaved  buffer is
dirtied and the VAO references it, it is reconstructed based on the
regular buffers. It is also reconstructed when the morph geometry keys
change, and reset if the chunk is rebuilt.

Performance improvement is similar to the earlier VAO patch, but it also
applies to morph geometry. There's also the benefit that we don't need to
switch back to the default VAO after every draw call.